convert the errors from missing model properties to typesafe soft-fails
Design decision: Because we can't have typeclasses even as type tags, we can
not enforce that the functions will be able to statically typecheck that the
given model can in fact return e.g. a genetics-relevant information. Failing
deep in the code because of missing implementation is not an option either.
This is the third choice -- if something's missing, we should just return a
correctly typed "empty" state. Arguably this is good also for converting and
writing models, where it reduces the amount of mixed possibilities for the
conversion.