avoid uncertainty creep
If there's complexity in data (stuff may be missing), it must be removed as
soon as possible in the process so that the core program logic may be stupid
and robust. I'm "removing" this complexity by using `maybemap` right at the
data parts that are missing, so that's perfectly obvious where the uncertainity
is (and, eventually, that we should aim for a nicer world where less nullable
fields in models are necessary... ;D).